+++ CHANGELOG.md
@@ -102,6 +102,7 @@ Semantic versioning in our case means:
 - Adds a new rule to detect duplicate conditions in `if`s and `elif`s, #2241
 - Adds a new rule to detect duplicate `case` pattens in `match`, #3206
 - Adds a new rule to find too many `match` subjects, #3201
+- Adds a new rule to detect too many `case` statements, #3202
 - Adds a new rule to find too complex `except` with too many exceptions
 - Adds a new rule to find too many `PEP695` type params
 - Adds a new rule to find useless ternary expressions, #1706